Inline Data Catalog TableUtils to its only use
